fix(MAX_RANGE): change parameter MAX_RANGE to long type

+++ src/main/java/cn/edu/tsinghua/iotdb/kairosdb/conf/ConfigDescriptor.java
@@ -52,7 +52,8 @@ public class ConfigDescriptor {
         config.DEBUG = Integer.parseInt(properties.getProperty("DEBUG", config.DEBUG + ""));
         config.CONNECTION_NUM = Integer.parseInt(properties.getProperty("CONNECTION_NUM", config.CONNECTION_NUM + ""));
         config.GROUP_BY_UNIT = Integer.parseInt(properties.getProperty("GROUP_BY_UNIT", config.GROUP_BY_UNIT + ""));
-        config.MAX_RANGE = Integer.parseInt(properties.getProperty("MAX_RANGE", config.MAX_RANGE + ""));
+        config.MAX_RANGE = Long.parseLong(properties.getProperty("MAX_RANGE", config.MAX_RANGE +
+            ""));
         config.PROFILE_INTERVAL = Integer.parseInt(properties.getProperty("PROFILE_INTERVAL", config.PROFILE_INTERVAL + ""));
         config.ENABLE_PROFILER = Boolean.parseBoolean(properties.getProperty("ENABLE_PROFILER",
             config.ENABLE_PROFILER + ""));
